Replaced AddUserButton with NoUserView
This commit removes the AddUserButton as it is no longer required.
Also when no user is logged in the GridView shows a new NoUserView
