+++ Translations/en.lproj/Localizable.strings
@@ -1069,6 +1069,9 @@
 /// Locked users
 "lockedUsers" = "Locked users";
 
+/// Login
+"login" = "Login";
+
 /// Logo
 "logo" = "Logo";
